libretro audio: fix resampler pitch/overflow, drop dead audio-buffer var
Sample-rate and quality fixes in the digital-sound load-time resampler
(file_vswap.cpp), plus removal of a dead SDL-era buffer variable.

- Pitch: advance the phase accumulator on samplefrac >= FRACUNIT (was '>').
  At exact equality the old test stalled an extra output sample on the same
  input position, consuming slightly fewer input samples than intended and
  detuning/slowing the sound.
- Pitch: the per-sound resample step was 'static const', so it was computed
  once from the first sound's rate and reused for every later sound; sounds
  with a different origSampleRate were resampled with the wrong step. Make it
  per-sound.
- Quality: the linear-interpolation product samplefrac*(next-cur) is up to
  ~4.3e9 and overflowed the int32 'fixed' before the shift. Compute it in
  64-bit. Also widen the per-voice multiply in the mixer templates to 64-bit;
  at full volume 32767*FRACUNIT is within one step of INT32_MAX, so the margin
  was unsafe.
- Remove param_audiobuffer: a leftover from the old SDL audio-callback buffer
  sizing. A libretro core does not own audio buffering (the frontend does), the
  variable was never read, and there is no under/overrun handling to keep.

8-bit sample paths (FORMAT_8BIT_LINEAR_UNSIGNED/SIGNED) are retained: they are
still reachable for raw 8-bit sound lumps in custom WAD/PK3 data loaded via
Mix_Chunk_Digital::loadSound (only VSWAP digis are pre-resampled to 16-bit).

+++ src/libretro/id_sd.cpp
@@ -623,8 +623,8 @@ mix_no_upscale(int16_t *result, const T *input, int num_samples, fixed leftmul, 
 	const int shift = FRACBITS - (16 - (int)sizeof(T) * 8);
 	int16_t *outptr = result;
 	for (int i = 0; i < num_samples; i++) {
-		int l = ((input[i] - bias) * leftmul) >> shift;
-		int r = ((input[i] - bias) * rightmul) >> shift;
+		int l = (int)(((long long)(input[i] - bias) * leftmul) >> shift);
+		int r = (int)(((long long)(input[i] - bias) * rightmul) >> shift);
 		mix_acc(outptr++, l);
 		mix_acc(outptr++, r);
 	}
@@ -637,8 +637,8 @@ mix_int_upscale(int16_t *result, const T *input, int num_samples, int upsample, 
 	const int shift = FRACBITS - (16 - (int)sizeof(T) * 8);
 	int16_t *outptr = result;
 	for (int i = 0; i < num_samples; i++) {
-		int l = ((input[i] - bias) * leftmul) >> shift;
-		int r = ((input[i] - bias) * rightmul) >> shift;
+		int l = (int)(((long long)(input[i] - bias) * leftmul) >> shift);
+		int r = (int)(((long long)(input[i] - bias) * rightmul) >> shift);
 		for (int j = 0 ; j < upsample; j++) {
 			mix_acc(outptr++, l);
 			mix_acc(outptr++, r);
@@ -653,8 +653,8 @@ mix_fractional_upscale(int16_t *result, int output_rate, const T *input, int inp
 	const int shift = FRACBITS - (16 - (int)sizeof(T) * 8);
 	int16_t *outptr = result;
 	for (int i = 0; i < num_samples; i++) {
-		int l = ((input[i] - bias) * leftmul) >> shift;
-		int r = ((input[i] - bias) * rightmul) >> shift;
+		int l = (int)(((long long)(input[i] - bias) * leftmul) >> shift);
+		int r = (int)(((long long)(input[i] - bias) * rightmul) >> shift);
 		while ((outptr - result) / 2 * (long long)input_rate < i * (long long) output_rate) {
 			mix_acc(outptr++, l);
 			mix_acc(outptr++, r);
